services/stream/requirements.txt
pytest
requests
boto3==1.35.90
boto3-stubs[essential]==1.35.90
psycopg[binary,pool]==3.3.3
sqlitedict==2.1.0
python-statemachine[diagrams]==2.5.0
osgb==1.2.0
